backend/requirements.txt
fastapi
psycopg2-binary
pydantic
python-dotenv
sqlalchemy
uvicorn
